Moving the code for the 3D road line strips to the Thing class. This will make the code more modular.
Made a change so that instead of cycling through 6 road markings, a new
road object will be created each time. This should be OK with garbage
collection? But worthy of looking into later. 

Added a new paramter to the Thing super constructor for the height of the
screen. This will be used to spwan the things at the horizon.